+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Buttons/ODSEmphasisButton.swift
@@ -40,11 +40,11 @@ public struct ODSButton: View {
     // MARK: Stored Properties
     // =======================
 
-    let text: LocalizedStringKey
-    let image: Image?
-    let emphasis: Emphasis
-    let variableWidth: Bool
-    let action: () -> Void
+    private let text: Text
+    private let image: Image?
+    private let emphasis: Emphasis
+    private let fullWidth: Bool
+    private let action: () -> Void
 
     // ==================
     // MARK: Initializers
@@ -56,20 +56,20 @@ public struct ODSButton: View {
     ///   - text: Text displayed in the button.
     ///   - image: Painter of the icon. If `nil`, no icon will be displayed.
     ///   - emphasis: Controls the style of the button. Use `ODSButton.Emphasis.highest` for an highlighted button style. To get a bordered button use `ODSButton.Emphasis.medium` and get a text only use `ODSButton.Emphasis.low`.
-    ///   - variableWidth: Defines the size of the button layout. Set to `true`, the size of the button is limited to the size of the text added by a padding round it. Set to `false` means button takes all available space horizontally.
+    ///   - fullWidth: Defines the size of the button layout. Set to `true` means button takes all available space horizontally. Set to `false`, the size of the button is limited to the size of the text added by a padding round it.
     ///   - action: Will be called when the user clicks the button.
     ///
     public init(
-        text: LocalizedStringKey,
+        text: Text,
         image: Image? = nil,
         emphasis: Emphasis,
-        variableWidth: Bool = true,
+        fullWidth: Bool = false,
         action: @escaping () -> Void
     ) {
         self.text = text
         self.image = image
         self.emphasis = emphasis
-        self.variableWidth = variableWidth
+        self.fullWidth = fullWidth
         self.action = action
     }
 
@@ -81,9 +81,9 @@ public struct ODSButton: View {
         Button {
             action()
         } label: {
-            ODSButtonContent(text: text, image: image, variableWidth: variableWidth)
+            ODSButtonContent(text, image: image)
         }
-        .modifier(ODSButtonStyleModifier(emphasis: emphasis))
+        .odsEmphasisButtonStyle(emphasis: emphasis, fullWidth: fullWidth)
     }
 }
 
@@ -96,13 +96,13 @@ struct ODSButton_Previews: PreviewProvider {
             ScrollView {
                 VStack {
                     ForEach(ODSButton.Emphasis.allCases, id: \.rawValue) { emphasis in
-                        ODSButton(text: LocalizedStringKey(emphasis.rawValue), emphasis: emphasis) {}
-                        ODSButton(text: LocalizedStringKey(emphasis.rawValue), emphasis: emphasis) {}.disabled(true)
-
-                        ODSButton(text: LocalizedStringKey(emphasis.rawValue),
+                        ODSButton(text: Text(emphasis.rawValue), emphasis: emphasis) {}
+                        ODSButton(text: Text(emphasis.rawValue), emphasis: emphasis) {}.disabled(true)
+                        ODSButton(text: Text(emphasis.rawValue), emphasis: emphasis) {}.disabled(true)
+                        ODSButton(text: Text(emphasis.rawValue),
                                   image: Image(systemName: "pencil.tip.crop.circle"),
                                   emphasis: emphasis) {}
-                        ODSButton(text: LocalizedStringKey(emphasis.rawValue),
+                        ODSButton(text: Text(emphasis.rawValue),
                                   image: Image(systemName: "pencil.tip.crop.circle"),
                                   emphasis: emphasis) {}.disabled(true)
                     }