+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Buttons/ODSFunctionalButton.swift
@@ -35,13 +35,11 @@ public struct ODSFunctionalButton: View {
     // MARK: Stored Properties
     // =======================
 
-    @Environment(\.theme) var theme
-
-    let text: LocalizedStringKey
-    let image: Image?
-    let style: ODSFunctionalButton.Style
-    let variableWidth: Bool
-    let action: () -> Void
+    private let text: Text
+    private let image: Image?
+    private let style: ODSFunctionalButton.Style
+    private let fullWidth: Bool
+    private let action: () -> Void
 
     // =================
     // MARK: Initializer
@@ -53,20 +51,20 @@ public struct ODSFunctionalButton: View {
     ///   - text: Text displayed in the button.
     ///   - image: Painter of the icon. If `nil`, no icon will be displayed.
     ///   - style: Controls the style of the button. To get a green/red buttons, you can use  `ODSFunctionalButton.Style.positive` or `ODSFunctionalButton.Style.negative`.
-    ///   - variableWidth: Defines the size of the button layout. Set to `true`, the size of the button is limited to the size of the text added by a padding round it. Set to `false` means button takes all available space horizontally.
+    ///   - fullWidth: Defines the size of the button layout. Set to `true` means button takes all available space horizontally. Set to `false`, the size of the button is limited to the size of the text added by a padding round it.
     ///   - action: Will be called when the user clicks the button.
     ///
     public init(
-        text: LocalizedStringKey,
+        text: Text,
         image: Image? = nil,
         style: ODSFunctionalButton.Style,
-        variableWidth: Bool = true,
+        fullWidth: Bool = false,
         action: @escaping () -> Void
     ) {
         self.text = text
         self.image = image
         self.style = style
-        self.variableWidth = variableWidth
+        self.fullWidth = fullWidth
         self.action = action
     }
 
@@ -78,26 +76,9 @@ public struct ODSFunctionalButton: View {
         Button {
             action()
         } label: {
-            ODSButtonContent(text: text, image: image, variableWidth: variableWidth)
+            ODSButtonContent(text, image: image)
         }
-        .buttonStyle(ODSShapedButtonStyle(shapeType: .filled, foregroundColor: foregroundColor, backgroundColor: backgroundColor))
-    }
-
-    // =====================
-    // MARK: Private Helpers
-    // =====================
-
-    private var foregroundColor: Color {
-        Color(UIColor.systemBackground)
-    }
-
-    private var backgroundColor: Color {
-        switch style {
-        case .negative:
-            return theme.componentColors.functionalNegative
-        case .positive:
-            return theme.componentColors.functionalPositive
-        }
+        .odsFunctionalButtonStyle(style: style, fullWidth: fullWidth)
     }
 }
 
@@ -107,14 +88,14 @@ struct ODSFunctionalButton_Previews: PreviewProvider {
         VStack(spacing: ODSSpacing.xl) {
             VStack {
                 Text("Without image")
-                ODSFunctionalButton(text: "Enabled", style: .negative) {}
-                ODSFunctionalButton(text: "Disabled", style: .negative) {}.disabled(true)
+                ODSFunctionalButton(text: Text("Enabled"), style: .negative) {}
+                ODSFunctionalButton(text: Text("Disabled") , style: .negative) {}.disabled(true)
             }
 
             VStack {
                 Text("Without image")
-                ODSFunctionalButton(text: "Enabled", image: Image(systemName: "pencil.tip"), style: ODSFunctionalButton.Style.negative) {}
-                ODSFunctionalButton(text: "Disabled", image: Image(systemName: "pencil.tip"), style: ODSFunctionalButton.Style.negative) {}.disabled(true)
+                ODSFunctionalButton(text: Text("Enabled"), image: Image(systemName: "pencil.tip"), style: ODSFunctionalButton.Style.negative) {}
+                ODSFunctionalButton(text: Text("Disabled"), image: Image(systemName: "pencil.tip"), style: ODSFunctionalButton.Style.negative) {}.disabled(true)
             }
         }
     }